requirements.txt
python-telegram-bot==21.6
openai
Pillow
requests